test: rename coverage_boost*_test.go files by feature for clarity
