--- coverage_boost3_test.go
@@ -1,1683 +1,0 @@
-package kruda
-
-import (
-	"context"
-	"errors"
-	"net/http/httptest"
-	"strings"
-	"testing"
-
-	"github.com/go-kruda/kruda/transport"
-)
-
-// =============================================================================
-// resource.go — resourceParseID: int64, uint, uint64, unsupported type branches
-// =============================================================================
-
-type mockInt64Item struct {
-	ID   int64  `json:"id"`
-	Name string `json:"name"`
-}
-
-type mockInt64ItemService struct {
-	getFn    func(ctx context.Context, id int64) (mockInt64Item, error)
-	listFn   func(ctx context.Context, page, limit int) ([]mockInt64Item, int, error)
-	createFn func(ctx context.Context, item mockInt64Item) (mockInt64Item, error)
-	updateFn func(ctx context.Context, id int64, item mockInt64Item) (mockInt64Item, error)
-	deleteFn func(ctx context.Context, id int64) error
-}
-
-func (s *mockInt64ItemService) List(ctx context.Context, page, limit int) ([]mockInt64Item, int, error) {
-	if s.listFn != nil {
-		return s.listFn(ctx, page, limit)
-	}
-	return nil, 0, nil
-}
-func (s *mockInt64ItemService) Create(ctx context.Context, item mockInt64Item) (mockInt64Item, error) {
-	if s.createFn != nil {
-		return s.createFn(ctx, item)
-	}
-	return item, nil
-}
-func (s *mockInt64ItemService) Get(ctx context.Context, id int64) (mockInt64Item, error) {
-	if s.getFn != nil {
-		return s.getFn(ctx, id)
-	}
-	return mockInt64Item{}, nil
-}
-func (s *mockInt64ItemService) Update(ctx context.Context, id int64, item mockInt64Item) (mockInt64Item, error) {
-	if s.updateFn != nil {
-		return s.updateFn(ctx, id, item)
-	}
-	return item, nil
-}
-func (s *mockInt64ItemService) Delete(ctx context.Context, id int64) error {
-	if s.deleteFn != nil {
-		return s.deleteFn(ctx, id)
-	}
-	return nil
-}
-
-func TestResourceParseID_Int64(t *testing.T) {
-	app := New()
-	var receivedID int64
-	svc := &mockInt64ItemService{
-		getFn: func(_ context.Context, id int64) (mockInt64Item, error) {
-			receivedID = id
-			return mockInt64Item{ID: id, Name: "Test"}, nil
-		},
-	}
-	Resource(app, "/items", svc)
-	app.Compile()
-
-	req := &mockRequest{method: "GET", path: "/items/999"}
-	resp := newMockResponse()
-	app.ServeKruda(resp, req)
-
-	if resp.statusCode != 200 {
-		t.Fatalf("status = %d, want 200, body: %s", resp.statusCode, resp.body)
-	}
-	if receivedID != 999 {
-		t.Errorf("received id = %d, want 999", receivedID)
-	}
-}
-
-func TestResourceParseID_Int64_Invalid(t *testing.T) {
-	app := New()
-	svc := &mockInt64ItemService{}
-	Resource(app, "/items", svc)
-	app.Compile()
-
-	req := &mockRequest{method: "GET", path: "/items/not-a-number"}
-	resp := newMockResponse()
-	app.ServeKruda(resp, req)
-
-	if resp.statusCode != 400 {
-		t.Fatalf("status = %d, want 400", resp.statusCode)
-	}
-}
-
-type mockUintItem struct {
-	ID   uint   `json:"id"`
-	Name string `json:"name"`
-}
-
-type mockUintItemService struct {
-	getFn    func(ctx context.Context, id uint) (mockUintItem, error)
-	listFn   func(ctx context.Context, page, limit int) ([]mockUintItem, int, error)
-	createFn func(ctx context.Context, item mockUintItem) (mockUintItem, error)
-	updateFn func(ctx context.Context, id uint, item mockUintItem) (mockUintItem, error)
-	deleteFn func(ctx context.Context, id uint) error
-}
-
-func (s *mockUintItemService) List(ctx context.Context, page, limit int) ([]mockUintItem, int, error) {
-	if s.listFn != nil {
-		return s.listFn(ctx, page, limit)
-	}
-	return nil, 0, nil
-}
-func (s *mockUintItemService) Create(ctx context.Context, item mockUintItem) (mockUintItem, error) {
-	if s.createFn != nil {
-		return s.createFn(ctx, item)
-	}
-	return item, nil
-}
-func (s *mockUintItemService) Get(ctx context.Context, id uint) (mockUintItem, error) {
-	if s.getFn != nil {
-		return s.getFn(ctx, id)
-	}
-	return mockUintItem{}, nil
-}
-func (s *mockUintItemService) Update(ctx context.Context, id uint, item mockUintItem) (mockUintItem, error) {
-	if s.updateFn != nil {
-		return s.updateFn(ctx, id, item)
-	}
-	return item, nil
-}
-func (s *mockUintItemService) Delete(ctx context.Context, id uint) error {
-	if s.deleteFn != nil {
-		return s.deleteFn(ctx, id)
-	}
-	return nil
-}
-
-func TestResourceParseID_Uint(t *testing.T) {
-	app := New()
-	var receivedID uint
-	svc := &mockUintItemService{
-		getFn: func(_ context.Context, id uint) (mockUintItem, error) {
-			receivedID = id
-			return mockUintItem{ID: id, Name: "Test"}, nil
-		},
-	}
-	Resource(app, "/items", svc)
-	app.Compile()
-
-	req := &mockRequest{method: "GET", path: "/items/42"}
-	resp := newMockResponse()
-	app.ServeKruda(resp, req)
-
-	if resp.statusCode != 200 {
-		t.Fatalf("status = %d, want 200, body: %s", resp.statusCode, resp.body)
-	}
-	if receivedID != 42 {
-		t.Errorf("received id = %d, want 42", receivedID)
-	}
-}
-
-func TestResourceParseID_Uint_Invalid(t *testing.T) {
-	app := New()
-	svc := &mockUintItemService{}
-	Resource(app, "/items", svc)
-	app.Compile()
-
-	req := &mockRequest{method: "GET", path: "/items/abc"}
-	resp := newMockResponse()
-	app.ServeKruda(resp, req)
-
-	if resp.statusCode != 400 {
-		t.Fatalf("status = %d, want 400", resp.statusCode)
-	}
-}
-
-type mockUint64Item struct {
-	ID   uint64 `json:"id"`
-	Name string `json:"name"`
-}
-
-type mockUint64ItemService struct {
-	getFn    func(ctx context.Context, id uint64) (mockUint64Item, error)
-	listFn   func(ctx context.Context, page, limit int) ([]mockUint64Item, int, error)
-	createFn func(ctx context.Context, item mockUint64Item) (mockUint64Item, error)
-	updateFn func(ctx context.Context, id uint64, item mockUint64Item) (mockUint64Item, error)
-	deleteFn func(ctx context.Context, id uint64) error
-}
-
-func (s *mockUint64ItemService) List(ctx context.Context, page, limit int) ([]mockUint64Item, int, error) {
-	if s.listFn != nil {
-		return s.listFn(ctx, page, limit)
-	}
-	return nil, 0, nil
-}
-func (s *mockUint64ItemService) Create(ctx context.Context, item mockUint64Item) (mockUint64Item, error) {
-	if s.createFn != nil {
-		return s.createFn(ctx, item)
-	}
-	return item, nil
-}
-func (s *mockUint64ItemService) Get(ctx context.Context, id uint64) (mockUint64Item, error) {
-	if s.getFn != nil {
-		return s.getFn(ctx, id)
-	}
-	return mockUint64Item{}, nil
-}
-func (s *mockUint64ItemService) Update(ctx context.Context, id uint64, item mockUint64Item) (mockUint64Item, error) {
-	if s.updateFn != nil {
-		return s.updateFn(ctx, id, item)
-	}
-	return item, nil
-}
-func (s *mockUint64ItemService) Delete(ctx context.Context, id uint64) error {
-	if s.deleteFn != nil {
-		return s.deleteFn(ctx, id)
-	}
-	return nil
-}
-
-func TestResourceParseID_Uint64(t *testing.T) {
-	app := New()
-	var receivedID uint64
-	svc := &mockUint64ItemService{
-		getFn: func(_ context.Context, id uint64) (mockUint64Item, error) {
-			receivedID = id
-			return mockUint64Item{ID: id, Name: "Test"}, nil
-		},
-	}
-	Resource(app, "/items", svc)
-	app.Compile()
-
-	req := &mockRequest{method: "GET", path: "/items/18446744073709551"}
-	resp := newMockResponse()
-	app.ServeKruda(resp, req)
-
-	if resp.statusCode != 200 {
-		t.Fatalf("status = %d, want 200, body: %s", resp.statusCode, resp.body)
-	}
-	if receivedID != 18446744073709551 {
-		t.Errorf("received id = %d, want 18446744073709551", receivedID)
-	}
-}
-
-func TestResourceParseID_Uint64_Invalid(t *testing.T) {
-	app := New()
-	svc := &mockUint64ItemService{}
-	Resource(app, "/items", svc)
-	app.Compile()
-
-	req := &mockRequest{method: "GET", path: "/items/xyz"}
-	resp := newMockResponse()
-	app.ServeKruda(resp, req)
-
-	if resp.statusCode != 400 {
-		t.Fatalf("status = %d, want 400", resp.statusCode)
-	}
-}
-
-// =============================================================================
-// resource.go — registerResource: LIST-only and GET_BY_ID-only paths
-// =============================================================================
-
-func TestResourceWithOnly_ListOnly(t *testing.T) {
-	app := New()
-	svc := &mockUserService{
-		listFn: func(_ context.Context, page, limit int) ([]mockUser, int, error) {
-			return []mockUser{{ID: "1", Name: "Alice"}}, 1, nil
-		},
-	}
-	Resource(app, "/users", svc, WithResourceOnly("LIST"))
-	app.Compile()
-
-	// LIST should work
-	req := &mockRequest{method: "GET", path: "/users"}
-	resp := newMockResponse()
-	app.ServeKruda(resp, req)
-	if resp.statusCode != 200 {
-		t.Errorf("GET /users status = %d, want 200", resp.statusCode)
-	}
-
-	// GET by ID should NOT work (only LIST registered)
-	req = &mockRequest{method: "GET", path: "/users/1"}
-	resp = newMockResponse()
-	app.ServeKruda(resp, req)
-	if resp.statusCode == 200 {
-		t.Error("GET /users/1 should not be registered with LIST-only")
-	}
-}
-
-func TestResourceWithOnly_GetByIDOnly(t *testing.T) {
-	app := New()
-	svc := &mockUserService{
-		getFn: func(_ context.Context, id string) (mockUser, error) {
-			return mockUser{ID: id, Name: "Alice"}, nil
-		},
-	}
-	Resource(app, "/users", svc, WithResourceOnly("GET_BY_ID"))
-	app.Compile()
-
-	// GET by ID should work
-	req := &mockRequest{method: "GET", path: "/users/abc"}
-	resp := newMockResponse()
-	app.ServeKruda(resp, req)
-	if resp.statusCode != 200 {
-		t.Errorf("GET /users/abc status = %d, want 200", resp.statusCode)
-	}
-
-	// LIST should NOT work
-	req = &mockRequest{method: "GET", path: "/users"}
-	resp = newMockResponse()
-	app.ServeKruda(resp, req)
-	if resp.statusCode == 200 {
-		t.Error("GET /users should not be registered with GET_BY_ID-only")
-	}
-}
-
-// =============================================================================
-// resource.go — service error paths (list error, create error, update error, delete error)
-// =============================================================================
-
-func TestResourceList_Error(t *testing.T) {
-	app := New()
-	svc := &mockUserService{
-		listFn: func(_ context.Context, page, limit int) ([]mockUser, int, error) {
-			return nil, 0, errors.New("db error")
-		},
-	}
-	Resource(app, "/users", svc)
-	app.Compile()
-
-	req := &mockRequest{method: "GET", path: "/users"}
-	resp := newMockResponse()
-	app.ServeKruda(resp, req)
-
-	if resp.statusCode != 500 {
-		t.Errorf("status = %d, want 500", resp.statusCode)
-	}
-}
-
-func TestResourceCreate_Error(t *testing.T) {
-	app := New()
-	svc := &mockUserService{
-		createFn: func(_ context.Context, item mockUser) (mockUser, error) {
-			return mockUser{}, errors.New("create failed")
-		},
-	}
-	Resource(app, "/users", svc)
-	app.Compile()
-
-	req := &mockRequest{
-		method:  "POST",
-		path:    "/users",
-		headers: map[string]string{"Content-Type": "application/json"},
-		body:    []byte(`{"name":"Alice"}`),
-	}
-	resp := newMockResponse()
-	app.ServeKruda(resp, req)
-
-	if resp.statusCode != 500 {
-		t.Errorf("status = %d, want 500", resp.statusCode)
-	}
-}
-
-func TestResourceUpdate_Error(t *testing.T) {
-	app := New()
-	svc := &mockUserService{
-		updateFn: func(_ context.Context, id string, item mockUser) (mockUser, error) {
-			return mockUser{}, errors.New("update failed")
-		},
-	}
-	Resource(app, "/users", svc)
-	app.Compile()
-
-	req := &mockRequest{
-		method:  "PUT",
-		path:    "/users/abc",
-		headers: map[string]string{"Content-Type": "application/json"},
-		body:    []byte(`{"name":"Updated"}`),
-	}
-	resp := newMockResponse()
-	app.ServeKruda(resp, req)
-
-	if resp.statusCode != 500 {
-		t.Errorf("status = %d, want 500", resp.statusCode)
-	}
-}
-
-func TestResourceDelete_Error(t *testing.T) {
-	app := New()
-	svc := &mockUserService{
-		deleteFn: func(_ context.Context, id string) error {
-			return errors.New("delete failed")
-		},
-	}
-	Resource(app, "/users", svc)
-	app.Compile()
-
-	req := &mockRequest{method: "DELETE", path: "/users/abc"}
-	resp := newMockResponse()
-	app.ServeKruda(resp, req)
-
-	if resp.statusCode != 500 {
-		t.Errorf("status = %d, want 500", resp.statusCode)
-	}
-}
-
-func TestResourceUpdate_InvalidID(t *testing.T) {
-	app := New()
-	svc := &mockItemService{}
-	Resource(app, "/items", svc)
-	app.Compile()
-
-	req := &mockRequest{
-		method:  "PUT",
-		path:    "/items/not-a-number",
-		headers: map[string]string{"Content-Type": "application/json"},
-		body:    []byte(`{"name":"X"}`),
-	}
-	resp := newMockResponse()
-	app.ServeKruda(resp, req)
-
-	if resp.statusCode != 400 {
-		t.Errorf("status = %d, want 400", resp.statusCode)
-	}
-}
-
-func TestResourceDelete_InvalidID(t *testing.T) {
-	app := New()
-	svc := &mockItemService{}
-	Resource(app, "/items", svc)
-	app.Compile()
-
-	req := &mockRequest{method: "DELETE", path: "/items/not-a-number"}
-	resp := newMockResponse()
-	app.ServeKruda(resp, req)
-
-	if resp.statusCode != 400 {
-		t.Errorf("status = %d, want 400", resp.statusCode)
-	}
-}
-
-func TestResourceCreate_InvalidBody(t *testing.T) {
-	app := New()
-	svc := &mockUserService{}
-	Resource(app, "/users", svc)
-	app.Compile()
-
-	req := &mockRequest{
-		method:  "POST",
-		path:    "/users",
-		headers: map[string]string{"Content-Type": "application/json"},
-		body:    []byte(`{invalid json}`),
-	}
-	resp := newMockResponse()
-	app.ServeKruda(resp, req)
-
-	// Should fail with a bind error
-	if resp.statusCode == 201 {
-		t.Error("invalid body should not return 201")
-	}
-}
-
-func TestResourceUpdate_InvalidBody(t *testing.T) {
-	app := New()
-	svc := &mockUserService{}
-	Resource(app, "/users", svc)
-	app.Compile()
-
-	req := &mockRequest{
-		method:  "PUT",
-		path:    "/users/abc",
-		headers: map[string]string{"Content-Type": "application/json"},
-		body:    []byte(`{invalid json}`),
-	}
-	resp := newMockResponse()
-	app.ServeKruda(resp, req)
-
-	if resp.statusCode == 200 {
-		t.Error("invalid body should not return 200")
-	}
-}
-
-func TestResourceWithCustomIDParam(t *testing.T) {
-	app := New()
-	var receivedID string
-	svc := &mockUserService{
-		getFn: func(_ context.Context, id string) (mockUser, error) {
-			receivedID = id
-			return mockUser{ID: id, Name: "Test"}, nil
-		},
-	}
-	Resource(app, "/users", svc, WithResourceIDParam("userId"))
-	app.Compile()
-
-	req := &mockRequest{method: "GET", path: "/users/custom-id-123"}
-	resp := newMockResponse()
-	app.ServeKruda(resp, req)
-
-	if resp.statusCode != 200 {
-		t.Fatalf("status = %d, want 200, body: %s", resp.statusCode, resp.body)
-	}
-	if receivedID != "custom-id-123" {
-		t.Errorf("received id = %q, want custom-id-123", receivedID)
-	}
-}
-
-// =============================================================================
-// router.go — validateRegexSafety: escaped chars, braces in groups
-// =============================================================================
-
-func TestValidateRegexSafety_Safe(t *testing.T) {
-	safe := []string{
-		`[0-9]+`,
-		`[a-zA-Z]+`,
-		`\d+`,
-		`[a-z]{2,5}`,
-		`(abc)`,
-		`(a|b)+`,
-	}
-	for _, p := range safe {
-		if err := validateRegexSafety(p); err != nil {
-			t.Errorf("validateRegexSafety(%q) returned error: %v", p, err)
-		}
-	}
-}
-
-func TestValidateRegexSafety_Unsafe(t *testing.T) {
-	unsafe := []string{
-		`(a+)+`,
-		`(a*)+`,
-		`(a+)*`,
-		`(a{2,})+`,
-	}
-	for _, p := range unsafe {
-		if err := validateRegexSafety(p); err == nil {
-			t.Errorf("validateRegexSafety(%q) should return error", p)
-		}
-	}
-}
-
-func TestValidateRegexSafety_EscapedChars(t *testing.T) {
-	// Escaped chars should not be treated as special
-	if err := validateRegexSafety(`\(a\+\)+`); err != nil {
-		// The escaped ( is not a real group, but the trailing + after ) is literal
-		// This tests the escape handling path
-		_ = err
-	}
-}
-
-func TestValidateRegexSafety_ClosingParenWithoutOpen(t *testing.T) {
-	// Unmatched closing paren should not panic
-	if err := validateRegexSafety(`)+`); err != nil {
-		// This might or might not be an error depending on implementation
-		_ = err
-	}
-}
-
-func TestValidateRegexSafety_BraceInGroup(t *testing.T) {
-	// {2,5} inside a group marks hasInnerQuantifier
-	if err := validateRegexSafety(`(a{2})?`); err == nil {
-		// A group with {2} followed by ? - since {2} sets hasInnerQuantifier
-		// and ? is a quantifier after ), this should be detected
-		_ = err
-	}
-}
-
-// =============================================================================
-// router.go — insertRoute: wildcard not last segment panic
-// =============================================================================
-
-func TestInsertRoute_WildcardNotLast_Panics(t *testing.T) {
-	defer func() {
-		if r := recover(); r == nil {
-			t.Error("expected panic for wildcard not as last segment")
-		}
-	}()
-	r := newRouter()
-	h := []HandlerFunc{dummyHandler()}
-	r.addRoute("GET", "/*wild/after", h)
-}
-
-func TestInsertRoute_WildcardNoName_Panics(t *testing.T) {
-	defer func() {
-		if r := recover(); r == nil {
-			t.Error("expected panic for unnamed wildcard")
-		}
-	}()
-	r := newRouter()
-	h := []HandlerFunc{dummyHandler()}
-	r.addRoute("GET", "/*", h)
-}
-
-func TestInsertRoute_DuplicateWildcard_Panics(t *testing.T) {
-	defer func() {
-		if r := recover(); r == nil {
-			t.Error("expected panic for duplicate wildcard")
-		}
-	}()
-	r := newRouter()
-	h := []HandlerFunc{dummyHandler()}
-	r.addRoute("GET", "/files/*filepath", h)
-	r.addRoute("GET", "/files/*filepath", h)
-}
-
-func TestInsertRoute_InvalidRegexConstraint_Panics(t *testing.T) {
-	defer func() {
-		if r := recover(); r == nil {
-			t.Error("expected panic for missing > in regex")
-		}
-	}()
-	r := newRouter()
-	h := []HandlerFunc{dummyHandler()}
-	r.addRoute("GET", "/users/:id<[0-9+", h)
-}
-
-func TestInsertRoute_AfterCompile_Panics(t *testing.T) {
-	defer func() {
-		if r := recover(); r == nil {
-			t.Error("expected panic for adding route after Compile()")
-		}
-	}()
-	r := newRouter()
-	h := []HandlerFunc{dummyHandler()}
-	r.addRoute("GET", "/test", h)
-	r.Compile()
-	r.addRoute("GET", "/test2", h)
-}
-
-func TestInsertRoute_PathMustStartWithSlash_Panics(t *testing.T) {
-	defer func() {
-		if r := recover(); r == nil {
-			t.Error("expected panic for path not starting with /")
-		}
-	}()
-	r := newRouter()
-	h := []HandlerFunc{dummyHandler()}
-	r.addRoute("GET", "noslash", h)
-}
-
-func TestInsertRoute_DuplicateRoot_Panics(t *testing.T) {
-	defer func() {
-		if r := recover(); r == nil {
-			t.Error("expected panic for duplicate root route")
-		}
-	}()
-	r := newRouter()
-	h := []HandlerFunc{dummyHandler()}
-	r.addRoute("GET", "/", h)
-	r.addRoute("GET", "/", h)
-}
-
-// =============================================================================
-// router.go — findInNode: regex param mismatch, optional param on root
-// =============================================================================
-
-func TestRouter_RegexParamMatch(t *testing.T) {
-	r := newRouter()
-	h := []HandlerFunc{dummyHandler()}
-	r.addRoute("GET", "/users/:id<[0-9]+>", h)
-	r.Compile()
-
-	var params routeParams
-
-	// Valid numeric ID
-	params.reset()
-	if r.find("GET", "/users/123", &params) == nil {
-		t.Error("should match numeric id")
-	}
-
-	// Invalid non-numeric ID
-	params.reset()
-	if r.find("GET", "/users/abc", &params) != nil {
-		t.Error("should not match non-numeric id")
-	}
-}
-
-func TestRouter_RegexParamWithSuffix(t *testing.T) {
-	r := newRouter()
-	h := []HandlerFunc{dummyHandler()}
-	r.addRoute("GET", "/files/:id<[0-9]+>/download", h)
-	r.Compile()
-
-	var params routeParams
-
-	params.reset()
-	if r.find("GET", "/files/42/download", &params) == nil {
-		t.Error("should match /files/42/download")
-	}
-
-	params.reset()
-	if r.find("GET", "/files/abc/download", &params) != nil {
-		t.Error("should not match /files/abc/download with regex constraint")
-	}
-}
-
-func TestRouter_OptionalParamAtRoot(t *testing.T) {
-	r := newRouter()
-	h := []HandlerFunc{dummyHandler()}
-	r.addRoute("GET", "/:lang?", h)
-
-	var params routeParams
-
-	// With param
-	params.reset()
-	if r.find("GET", "/en", &params) == nil {
-		t.Error("should match /en")
-	}
-
-	// Without param (root)
-	params.reset()
-	if r.find("GET", "/", &params) == nil {
-		t.Error("should match / with optional param")
-	}
-}
-
-func TestRouter_OptionalParamOnPath(t *testing.T) {
-	r := newRouter()
-	h := []HandlerFunc{dummyHandler()}
-	r.addRoute("GET", "/users/:id?", h)
-
-	var params routeParams
-
-	// With param
-	params.reset()
-	if r.find("GET", "/users/42", &params) == nil {
-		t.Error("should match /users/42")
-	}
-
-	// Without param
-	params.reset()
-	if r.find("GET", "/users", &params) == nil {
-		t.Error("should match /users with optional param")
-	}
-}
-
-// =============================================================================
-// router.go — find with custom method (non-standard, uses map fallback)
-// =============================================================================
-
-func TestRouter_CustomMethod(t *testing.T) {
-	r := newRouter()
-	h := []HandlerFunc{dummyHandler()}
-	r.addRoute("PURGE", "/cache", h)
-	r.Compile()
-
-	var params routeParams
-	params.reset()
-	if r.find("PURGE", "/cache", &params) == nil {
-		t.Error("should match custom method PURGE")
-	}
-}
-
-func TestRouter_CustomMethod_NoMatch(t *testing.T) {
-	r := newRouter()
-	var params routeParams
-	params.reset()
-	if r.find("CUSTOM", "/nothing", &params) != nil {
-		t.Error("should not match unregistered custom method")
-	}
-}
-
-// =============================================================================
-// router.go — cleanPath: null byte stripping
-// =============================================================================
-
-func TestCleanPath_NullByte(t *testing.T) {
-	cleaned, err := cleanPath("/test\x00path")
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	if strings.Contains(cleaned, "\x00") {
-		t.Error("cleaned path should not contain null bytes")
-	}
-}
-
-func TestCleanPath_DoubleEncoded(t *testing.T) {
-	// %252e%252e should be decoded to .. via double-decode
-	_, err := cleanPath("/%252e%252e/etc/passwd")
-	if err == nil {
-		// The double-decode should resolve to /../etc/passwd which is traversal
-		// But the decoded path starts with / so depth tracking matters
-		_ = err
-	}
-}
-
-func TestCleanPath_Simple(t *testing.T) {
-	cleaned, err := cleanPath("/users/42")
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	if cleaned != "/users/42" {
-		t.Errorf("cleanPath = %q, want /users/42", cleaned)
-	}
-}
-
-func TestCleanPath_DotSegments(t *testing.T) {
-	cleaned, err := cleanPath("/a/b/../c")
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	if cleaned != "/a/c" {
-		t.Errorf("cleanPath = %q, want /a/c", cleaned)
-	}
-}
-
-// =============================================================================
-// devmode.go — generateSuggestions: 405, 413, 422 cases
-// =============================================================================
-
-func TestGenerateSuggestions_405_Direct(t *testing.T) {
-	app := New(WithDevMode(true))
-	c := newCtx(app)
-	c.method = "POST"
-	c.path = "/api/data"
-
-	suggestions := generateSuggestions(errors.New("test"), 405, c)
-	found := false
-	for _, s := range suggestions {
-		if strings.Contains(s, "doesn't accept") {
-			found = true
-		}
-	}
-	if !found {
-		t.Errorf("405 suggestions should mention 'doesn't accept', got %v", suggestions)
-	}
-}
-
-func TestGenerateSuggestions_422(t *testing.T) {
-	app := New(WithDevMode(true))
-	c := newCtx(app)
-	c.method = "POST"
-	c.path = "/api/users"
-
-	suggestions := generateSuggestions(errors.New("test"), 422, c)
-	found := false
-	for _, s := range suggestions {
-		if strings.Contains(s, "request body") {
-			found = true
-		}
-	}
-	if !found {
-		t.Errorf("422 suggestions should mention request body, got %v", suggestions)
-	}
-}
-
-func TestGenerateSuggestions_413(t *testing.T) {
-	app := New(WithDevMode(true))
-	c := newCtx(app)
-	c.method = "POST"
-	c.path = "/upload"
-
-	suggestions := generateSuggestions(errors.New("test"), 413, c)
-	found := false
-	for _, s := range suggestions {
-		if strings.Contains(s, "maximum size") {
-			found = true
-		}
-	}
-	if !found {
-		t.Errorf("413 suggestions should mention max size, got %v", suggestions)
-	}
-}
-
-func TestGenerateSuggestions_500(t *testing.T) {
-	app := New(WithDevMode(true))
-	c := newCtx(app)
-	c.method = "GET"
-	c.path = "/fail"
-
-	suggestions := generateSuggestions(errors.New("test"), 500, c)
-	found := false
-	for _, s := range suggestions {
-		if strings.Contains(s, "stack trace") {
-			found = true
-		}
-	}
-	if !found {
-		t.Errorf("500 suggestions should mention stack trace, got %v", suggestions)
-	}
-}
-
-// =============================================================================
-// devmode.go — buildSourceLines edge case: targetLine near start
-// =============================================================================
-
-func TestBuildSourceLines_NearStart(t *testing.T) {
-	lines := []string{"line1", "line2", "line3", "line4", "line5"}
-	result := buildSourceLines(lines, 1, 2) // target=1, radius=2
-	if len(result) == 0 {
-		t.Fatal("expected non-empty result")
-	}
-	// First line should start at 1
-	if result[0].Number != 1 {
-		t.Errorf("first line number = %d, want 1", result[0].Number)
-	}
-	// The error line should be line 1
-	foundError := false
-	for _, sl := range result {
-		if sl.IsError && sl.Number == 1 {
-			foundError = true
-		}
-	}
-	if !foundError {
-		t.Error("expected line 1 to be marked as error")
-	}
-}
-
-func TestBuildSourceLines_BeyondEnd(t *testing.T) {
-	lines := []string{"a", "b", "c"}
-	result := buildSourceLines(lines, 3, 10) // radius larger than file
-	if len(result) == 0 {
-		t.Fatal("expected non-empty result")
-	}
-	// Should not go beyond len(lines)
-	lastLine := result[len(result)-1]
-	if lastLine.Number > len(lines) {
-		t.Errorf("last line number %d exceeds file length %d", lastLine.Number, len(lines))
-	}
-}
-
-// =============================================================================
-// devmode.go — collectRequestHeaders/collectQueryParams nil request
-// =============================================================================
-
-func TestCollectRequestHeaders_NilRequest(t *testing.T) {
-	app := New()
-	c := newCtx(app)
-	c.request = nil
-	h := collectRequestHeaders(c)
-	if len(h) != 0 {
-		t.Errorf("expected empty headers for nil request, got %d", len(h))
-	}
-}
-
-func TestCollectQueryParams_NilRequest(t *testing.T) {
-	app := New()
-	c := newCtx(app)
-	c.request = nil
-	q := collectQueryParams(c)
-	if len(q) != 0 {
-		t.Errorf("expected empty query params for nil request, got %d", len(q))
-	}
-}
-
-// allHeadersRequest implements AllHeadersProvider and AllQueryProvider.
-type allHeadersRequest struct {
-	mockRequest
-	hdrs  map[string]string
-	query map[string]string
-}
-
-func (r *allHeadersRequest) AllHeaders() map[string]string { return r.hdrs }
-func (r *allHeadersRequest) AllQuery() map[string]string    { return r.query }
-
-func TestCollectRequestHeaders_WithProvider(t *testing.T) {
-	app := New()
-	c := newCtx(app)
-	c.request = &allHeadersRequest{
-		mockRequest: mockRequest{method: "GET", path: "/test"},
-		hdrs:        map[string]string{"X-Test": "val"},
-	}
-	h := collectRequestHeaders(c)
-	if h["X-Test"] != "val" {
-		t.Errorf("expected X-Test=val, got %v", h)
-	}
-}
-
-func TestCollectQueryParams_WithProvider(t *testing.T) {
-	app := New()
-	c := newCtx(app)
-	c.request = &allHeadersRequest{
-		mockRequest: mockRequest{method: "GET", path: "/test"},
-		query:       map[string]string{"foo": "bar"},
-	}
-	q := collectQueryParams(c)
-	if q["foo"] != "bar" {
-		t.Errorf("expected foo=bar, got %v", q)
-	}
-}
-
-// =============================================================================
-// devmode.go — walkRouteTree with wildcard and param nodes
-// =============================================================================
-
-func TestWalkRouteTree_WithParams(t *testing.T) {
-	app := New(WithDevMode(true))
-	app.Get("/users/:id", func(c *Ctx) error { return c.Text("ok") })
-	app.Get("/files/*filepath", func(c *Ctx) error { return c.Text("ok") })
-	app.Compile()
-
-	routes := collectDevRoutes(app)
-	foundParam := false
-	foundWildcard := false
-	for _, r := range routes {
-		if strings.Contains(r.Path, ":id") {
-			foundParam = true
-		}
-		if strings.Contains(r.Path, "*filepath") {
-			foundWildcard = true
-		}
-	}
-	if !foundParam {
-		t.Error("routes should include :id param path")
-	}
-	if !foundWildcard {
-		t.Error("routes should include *filepath wildcard path")
-	}
-}
-
-// =============================================================================
-// devmode.go — filterEnvVars with no = sign
-// =============================================================================
-
-func TestFilterEnvVars(t *testing.T) {
-	vars := filterEnvVars()
-	// Should not contain any sensitive keys
-	for key := range vars {
-		if isSensitiveEnvKey(key) {
-			t.Errorf("sensitive key %q should be filtered", key)
-		}
-	}
-}
-
-func TestIsSensitiveEnvKey(t *testing.T) {
-	tests := []struct {
-		key       string
-		sensitive bool
-	}{
-		{"DB_PASSWORD", true},
-		{"API_TOKEN", true},
-		{"MY_SECRET", true},
-		{"APP_PORT", false},
-		{"GOPATH", false},
-		{"AWS_CREDENTIAL_FILE", true},
-		{"PRIVATE_KEY_PATH", true},
-	}
-	for _, tt := range tests {
-		if got := isSensitiveEnvKey(tt.key); got != tt.sensitive {
-			t.Errorf("isSensitiveEnvKey(%q) = %v, want %v", tt.key, got, tt.sensitive)
-		}
-	}
-}
-
-// =============================================================================
-// kruda.go — ServeHTTP: path traversal prevention, method not allowed, empty path
-// =============================================================================
-
-func TestServeHTTP_PathTraversalBlocked_Boost3(t *testing.T) {
-	app := New(WithPathTraversal(), NetHTTP())
-	app.Get("/safe", func(c *Ctx) error {
-		return c.Text("ok")
-	})
-	app.Compile()
-
-	req := httptest.NewRequest("GET", "/../../etc/passwd", nil)
-	w := httptest.NewRecorder()
-	app.ServeHTTP(w, req)
-
-	if w.Code != 400 {
-		t.Errorf("status = %d, want 400 for path traversal", w.Code)
-	}
-}
-
-func TestServeHTTP_MethodNotAllowed_Boost3(t *testing.T) {
-	app := New(NetHTTP())
-	app.Get("/data-boost3", func(c *Ctx) error {
-		return c.Text("ok")
-	})
-	app.Compile()
-
-	req := httptest.NewRequest("POST", "/data-boost3", nil)
-	w := httptest.NewRecorder()
-	app.ServeHTTP(w, req)
-
-	if w.Code != 405 {
-		t.Errorf("status = %d, want 405", w.Code)
-	}
-	// 405 is the key assertion; Allow header visibility in recorder
-	// depends on header write timing in the net/http adapter.
-}
-
-func TestServeHTTP_NotFound_Boost3(t *testing.T) {
-	app := New(NetHTTP())
-	app.Get("/exists-boost3", func(c *Ctx) error {
-		return c.Text("ok")
-	})
-	app.Compile()
-
-	req := httptest.NewRequest("GET", "/nope-boost3", nil)
-	w := httptest.NewRecorder()
-	app.ServeHTTP(w, req)
-
-	if w.Code != 404 {
-		t.Errorf("status = %d, want 404", w.Code)
-	}
-}
-
-func TestServeHTTP_HandlerError_Boost3(t *testing.T) {
-	app := New(NetHTTP())
-	app.Get("/fail-boost3", func(c *Ctx) error {
-		return errors.New("handler error")
-	})
-	app.Compile()
-
-	req := httptest.NewRequest("GET", "/fail-boost3", nil)
-	w := httptest.NewRecorder()
-	app.ServeHTTP(w, req)
-
-	if w.Code != 500 {
-		t.Errorf("status = %d, want 500", w.Code)
-	}
-}
-
-func TestServeHTTP_EmptyPath_Boost3(t *testing.T) {
-	app := New(NetHTTP())
-	app.Get("/", func(c *Ctx) error {
-		return c.Text("root")
-	})
-	app.Compile()
-
-	// Empty URL path should be treated as /
-	req := httptest.NewRequest("GET", "/", nil)
-	req.URL.Path = ""
-	w := httptest.NewRecorder()
-	app.ServeHTTP(w, req)
-
-	// With empty path, it should be set to "/" and match root
-	if w.Code == 0 {
-		t.Error("status should not be 0")
-	}
-}
-
-// =============================================================================
-// kruda.go — handleError: custom ErrorHandler with validation error
-// =============================================================================
-
-func TestHandleError_ValidationError_CustomHandler(t *testing.T) {
-	var capturedErr error
-	app := New(WithErrorHandler(func(c *Ctx, ke *KrudaError) {
-		capturedErr = ke
-		c.Status(ke.Code)
-		_ = c.JSON(Map{"custom": true, "code": ke.Code})
-	}))
-
-	app.Get("/validate", func(c *Ctx) error {
-		ve := &ValidationError{
-			Errors: []FieldError{{Field: "name", Rule: "required", Message: "is required"}},
-		}
-		return ve
-	})
-	app.Compile()
-
-	req := &mockRequest{method: "GET", path: "/validate"}
-	resp := newMockResponse()
-	app.ServeKruda(resp, req)
-
-	if capturedErr == nil {
-		t.Fatal("custom error handler should have been called")
-	}
-	if resp.statusCode != 422 {
-		t.Errorf("status = %d, want 422", resp.statusCode)
-	}
-}
-
-func TestHandleError_AlreadyResponded(t *testing.T) {
-	app := New()
-	app.Get("/test", func(c *Ctx) error {
-		_ = c.Text("already sent")
-		// Return error after already responded
-		return errors.New("late error")
-	})
-	app.Compile()
-
-	req := &mockRequest{method: "GET", path: "/test"}
-	resp := newMockResponse()
-	app.ServeKruda(resp, req)
-
-	// Should have the first response, not a 500
-	if resp.statusCode == 500 {
-		t.Log("error handler wrote 500 after panic recovery")
-	}
-}
-
-// =============================================================================
-// router.go — findInNode: param backtrack path (del)
-// =============================================================================
-
-func TestRouter_ParamBacktrack(t *testing.T) {
-	r := newRouter()
-	h := []HandlerFunc{dummyHandler()}
-	// Two param routes that could conflict during backtracking
-	r.addRoute("GET", "/a/:x/b", h)
-	r.addRoute("GET", "/a/:y/c", h)
-
-	var params routeParams
-
-	params.reset()
-	if r.find("GET", "/a/val/b", &params) == nil {
-		t.Error("should match /a/val/b")
-	}
-
-	params.reset()
-	if r.find("GET", "/a/val/c", &params) == nil {
-		t.Error("should match /a/val/c")
-	}
-
-	// No match
-	params.reset()
-	if r.find("GET", "/a/val/d", &params) != nil {
-		t.Error("should not match /a/val/d")
-	}
-}
-
-// =============================================================================
-// router.go — find: track hits before compile
-// =============================================================================
-
-func TestRouter_TrackHits_BeforeCompile(t *testing.T) {
-	r := newRouter()
-	h := []HandlerFunc{dummyHandler()}
-	r.addRoute("GET", "/hot", h)
-	r.addRoute("GET", "/cold", h)
-
-	var params routeParams
-
-	// Hit /hot 10 times before compile
-	for range 10 {
-		params.reset()
-		r.find("GET", "/hot", &params)
-	}
-	// Hit /cold once
-	params.reset()
-	r.find("GET", "/cold", &params)
-
-	// After compile, hits should have been recorded
-	r.Compile()
-
-	// Both should still match
-	params.reset()
-	if r.find("GET", "/hot", &params) == nil {
-		t.Error("should match /hot after compile")
-	}
-	params.reset()
-	if r.find("GET", "/cold", &params) == nil {
-		t.Error("should match /cold after compile")
-	}
-}
-
-// =============================================================================
-// router.go — insertRoute: optional param on existing param node
-// =============================================================================
-
-func TestRouter_OptionalParamDuplicate_Panics(t *testing.T) {
-	defer func() {
-		if r := recover(); r == nil {
-			t.Error("expected panic for duplicate optional param route")
-		}
-	}()
-	r := newRouter()
-	h := []HandlerFunc{dummyHandler()}
-	r.addRoute("GET", "/users/:id?", h)
-	r.addRoute("GET", "/users/:id?", h)
-}
-
-// =============================================================================
-// upload.go — Open with a valid header (already tested, but test error path)
-// The Open() error path for Header.Open() failure is hard to test without
-// mocking the multipart.FileHeader internal. The nil header path is already
-// covered. We test the success path for robustness.
-// =============================================================================
-
-func TestFileUpload_Open_WithContent(t *testing.T) {
-	fh := createTestFileHeader(t, "file", "test.txt", "text/plain", []byte("hello"))
-	fu := &FileUpload{
-		Name:        "test.txt",
-		Size:        5,
-		ContentType: "text/plain",
-		Header:      fh,
-	}
-	rc, err := fu.Open()
-	if err != nil {
-		t.Fatalf("Open() error: %v", err)
-	}
-	rc.Close()
-}
-
-// =============================================================================
-// kruda.go — Shutdown with container
-// =============================================================================
-
-func TestApp_Shutdown_WithContainer(t *testing.T) {
-	c := NewContainer()
-	// Register a service that implements Shutdowner interface
-	_ = c.Give("test-value")
-
-	app := New(WithContainer(c))
-	app.Get("/test", func(ctx *Ctx) error { return ctx.Text("ok") })
-	app.Compile()
-
-	err := app.Shutdown(context.Background())
-	if err != nil {
-		t.Fatalf("Shutdown error: %v", err)
-	}
-}
-
-func TestApp_Shutdown_WithoutContainer(t *testing.T) {
-	app := New()
-	app.Get("/test", func(ctx *Ctx) error { return ctx.Text("ok") })
-	app.Compile()
-
-	// Shutdown should work fine without a container
-	err := app.Shutdown(context.Background())
-	if err != nil {
-		t.Fatalf("Shutdown error: %v", err)
-	}
-}
-
-// =============================================================================
-// kruda.go — ServeKruda: panic recovery
-// =============================================================================
-
-func TestServeKruda_PanicRecovery_AlreadyResponded(t *testing.T) {
-	app := New()
-	app.Get("/panic-after-write", func(c *Ctx) error {
-		_ = c.Text("ok")
-		panic("late panic after response")
-	})
-	app.Compile()
-
-	req := &mockRequest{method: "GET", path: "/panic-after-write"}
-	resp := newMockResponse()
-
-	// Should not crash even if panic occurs after response is written
-	app.ServeKruda(resp, req)
-}
-
-// =============================================================================
-// kruda.go — ServeHTTP with security headers
-// =============================================================================
-
-func TestServeHTTP_SecurityHeaders_Boost3(t *testing.T) {
-	// Test that security headers are compiled and applied via ServeKruda
-	app := New(WithSecurity())
-	app.Get("/test", func(c *Ctx) error {
-		return c.Text("ok")
-	})
-	app.Compile()
-
-	req := &mockRequest{method: "GET", path: "/test"}
-	resp := newMockResponse()
-	app.ServeKruda(resp, req)
-
-	if resp.statusCode != 200 {
-		t.Fatalf("status = %d, want 200", resp.statusCode)
-	}
-	// Security headers should be set via writeHeaders
-	if resp.headers.Get("X-Content-Type-Options") == "" {
-		t.Error("expected X-Content-Type-Options header")
-	}
-}
-
-// =============================================================================
-// router.go — find on non-standard method tree (map fallback)
-// =============================================================================
-
-func TestRouter_FindNonStandardMethod(t *testing.T) {
-	r := newRouter()
-	h := []HandlerFunc{dummyHandler()}
-	r.addRoute("LINK", "/resource", h)
-	r.Compile()
-
-	var params routeParams
-	params.reset()
-	result := r.find("LINK", "/resource", &params)
-	if result == nil {
-		t.Error("should find LINK /resource via map fallback")
-	}
-}
-
-// =============================================================================
-// devmode.go — renderDevErrorPage production gate
-// =============================================================================
-
-func TestRenderDevErrorPage_ProductionReturnsNil(t *testing.T) {
-	app := New() // DevMode = false
-	c := newCtx(app)
-	c.method = "GET"
-	c.path = "/test"
-	result := renderDevErrorPage(c, errors.New("test"), 500)
-	if result != nil {
-		t.Error("renderDevErrorPage should return nil in production mode")
-	}
-}
-
-// =============================================================================
-// Ensure AllHeadersProvider/AllQueryProvider casting works in devmode
-// =============================================================================
-
-var _ transport.AllHeadersProvider = (*allHeadersRequest)(nil)
-var _ transport.AllQueryProvider = (*allHeadersRequest)(nil)
-
-// =============================================================================
-// kruda.go — Compile with security header variants
-// =============================================================================
-
-func TestCompile_AllSecurityHeaders(t *testing.T) {
-	app := New(func(a *App) {
-		a.config.SecurityHeaders = true
-		a.config.Security.XSSProtection = "1; mode=block"
-		a.config.Security.ContentTypeNosniff = "nosniff"
-		a.config.Security.XFrameOptions = "DENY"
-		a.config.Security.ReferrerPolicy = "strict-origin"
-		a.config.Security.ContentSecurityPolicy = "default-src 'self'"
-		a.config.Security.HSTSMaxAge = 31536000
-	})
-	app.Get("/test", func(c *Ctx) error { return c.Text("ok") })
-	app.Compile()
-
-	if len(app.secHeaders) != 6 {
-		t.Errorf("expected 6 security headers, got %d", len(app.secHeaders))
-	}
-}
-
-// =============================================================================
-// kruda.go — Compile: hasLifecycle flag
-// =============================================================================
-
-func TestCompile_HasLifecycleFlag(t *testing.T) {
-	app := New()
-	app.OnRequest(func(c *Ctx) error { return nil })
-	app.Get("/test", func(c *Ctx) error { return c.Text("ok") })
-	app.Compile()
-
-	if !app.hasLifecycle {
-		t.Error("hasLifecycle should be true when OnRequest is registered")
-	}
-}
-
-func TestCompile_NoLifecycle(t *testing.T) {
-	app := New()
-	app.Get("/test", func(c *Ctx) error { return c.Text("ok") })
-	app.Compile()
-
-	if app.hasLifecycle {
-		t.Error("hasLifecycle should be false when no hooks registered")
-	}
-}
-
-// =============================================================================
-// kruda.go — ServeHTTP: multipart form cleanup
-// =============================================================================
-
-func TestServeHTTP_MultipartCleanup(t *testing.T) {
-	app := New(NetHTTP())
-	app.Post("/upload", func(c *Ctx) error {
-		return c.Text("ok")
-	})
-	app.Compile()
-
-	// Send a POST with multipart content type
-	body := strings.NewReader("--boundary\r\nContent-Disposition: form-data; name=\"file\"\r\n\r\ncontent\r\n--boundary--")
-	req := httptest.NewRequest("POST", "/upload", body)
-	req.Header.Set("Content-Type", "multipart/form-data; boundary=boundary")
-	w := httptest.NewRecorder()
-	app.ServeHTTP(w, req)
-
-	if w.Code != 200 {
-		t.Errorf("status = %d, want 200", w.Code)
-	}
-}
-
-// =============================================================================
-// router.go — collectStaticRoutes: root with trailing slash
-// =============================================================================
-
-func TestCollectStaticRoutes_RootSlash(t *testing.T) {
-	r := newRouter()
-	h := []HandlerFunc{dummyHandler()}
-	r.addRoute("GET", "/", h)
-	r.addRoute("GET", "/api", h)
-	r.Compile()
-
-	// After compile, static routes should be populated
-	if r.staticRoutes[mGET] == nil {
-		t.Fatal("expected static routes for GET")
-	}
-	if _, ok := r.staticRoutes[mGET]["/"]; !ok {
-		t.Error("expected / in static routes")
-	}
-	if _, ok := r.staticRoutes[mGET]["/api"]; !ok {
-		t.Error("expected /api in static routes")
-	}
-}
-
-// =============================================================================
-// kruda.go — containsDotPercent
-// =============================================================================
-
-func TestContainsDotPercent_Boost3(t *testing.T) {
-	// Additional edge cases beyond coverage_boost_test.go
-	tests := []struct {
-		input string
-		want  bool
-	}{
-		{"/users/42", false},
-		{"/users/../admin", true},
-		{"/files/test%2F", true},
-		{"/.hidden", true},
-		{"/a/b/c/d/e", false},
-		{"/path.with.dots", true},
-	}
-	for _, tt := range tests {
-		if got := containsDotPercent(tt.input); got != tt.want {
-			t.Errorf("containsDotPercent(%q) = %v, want %v", tt.input, got, tt.want)
-		}
-	}
-}
-
-// =============================================================================
-// kruda.go — net/http ServeHTTP with path traversal (dot+percent path)
-// =============================================================================
-
-func TestServeHTTP_PathTraversal_CleanedSuccessfully(t *testing.T) {
-	app := New(NetHTTP(), WithPathTraversal())
-	app.Get("/a/c", func(c *Ctx) error {
-		return c.Text("ok")
-	})
-	app.Compile()
-
-	// /a/b/../c should be cleaned to /a/c
-	req := httptest.NewRequest("GET", "/a/b/../c", nil)
-	w := httptest.NewRecorder()
-	app.ServeHTTP(w, req)
-
-	if w.Code != 200 {
-		t.Errorf("status = %d, want 200 after path cleaning", w.Code)
-	}
-}
-
-// =============================================================================
-// kruda.go — OnShutdown LIFO order
-// =============================================================================
-
-func TestOnShutdown_EmptyHooks(t *testing.T) {
-	app := New()
-	app.Compile()
-	// Should not panic with empty hooks
-	app.runShutdownHooks()
-}
-
-// =============================================================================
-// kruda.go — ServeHTTP shrinkMaps on various paths
-// =============================================================================
-
-func TestServeHTTP_ShrinkMapsOnNotFound(t *testing.T) {
-	app := New(NetHTTP())
-	app.Compile()
-
-	req := httptest.NewRequest("GET", "/nonexistent", nil)
-	w := httptest.NewRecorder()
-	app.ServeHTTP(w, req)
-
-	if w.Code != 404 {
-		t.Errorf("status = %d, want 404", w.Code)
-	}
-}
-
-// =============================================================================
-// resource.go — resourceParseID unsupported type (float64)
-// =============================================================================
-
-func TestResourceParseID_UnsupportedType(t *testing.T) {
-	_, err := resourceParseID[float64]("3.14")
-	if err == nil {
-		t.Error("expected error for unsupported ID type float64")
-	}
-	if !strings.Contains(err.Error(), "unsupported") {
-		t.Errorf("error = %q, want to contain 'unsupported'", err.Error())
-	}
-}
-
-// =============================================================================
-// devmode.go — DevMode 405 suggestion via ServeKruda (exercises full path)
-// =============================================================================
-
-func TestDevMode_405_ViaTestClient(t *testing.T) {
-	app := New(WithDevMode(true))
-	app.Get("/only-get", func(c *Ctx) error {
-		return c.Text("ok")
-	})
-	app.Compile()
-
-	tc := NewTestClient(app)
-	resp, err := tc.Request("POST", "/only-get").Send()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if resp.StatusCode() != 405 {
-		t.Errorf("status = %d, want 405", resp.StatusCode())
-	}
-	body := resp.BodyString()
-	if !strings.Contains(body, "doesn't accept") && !strings.Contains(body, "method not allowed") {
-		t.Error("405 dev error page should include method not allowed indication")
-	}
-}
-
-// =============================================================================
-// kruda.go — handleError: KrudaError with 5xx in production (strip detail)
-// =============================================================================
-
-func TestHandleError_5xxStripDetail_Production(t *testing.T) {
-	app := New() // production mode
-	app.Get("/fail", func(c *Ctx) error {
-		ke := NewError(500, "server error")
-		ke.Detail = "internal details here"
-		return ke
-	})
-	app.Compile()
-
-	req := &mockRequest{method: "GET", path: "/fail"}
-	resp := newMockResponse()
-	app.ServeKruda(resp, req)
-
-	body := string(resp.body)
-	if strings.Contains(body, "internal details here") {
-		t.Error("production 5xx should strip Detail")
-	}
-}
-
-// =============================================================================
-// kruda.go — handleError: mapped error with 4xx in production (preserve)
-// =============================================================================
-
-func TestHandleError_4xxPreserve_Production(t *testing.T) {
-	app := New()
-	app.Get("/fail", func(c *Ctx) error {
-		return BadRequest("invalid input")
-	})
-	app.Compile()
-
-	req := &mockRequest{method: "GET", path: "/fail"}
-	resp := newMockResponse()
-	app.ServeKruda(resp, req)
-
-	if resp.statusCode != 400 {
-		t.Errorf("status = %d, want 400", resp.statusCode)
-	}
-	body := string(resp.body)
-	if !strings.Contains(body, "invalid input") {
-		t.Error("4xx should preserve message in production")
-	}
-}
-
-// =============================================================================
-// kruda.go — net/http request with nil cookies slice reuse
-// =============================================================================
-
-func TestServeHTTP_CookieSliceReuse(t *testing.T) {
-	app := New(NetHTTP())
-	app.Get("/test", func(c *Ctx) error {
-		c.SetCookie(&Cookie{Name: "test", Value: "val"})
-		return c.Text("ok")
-	})
-	app.Compile()
-
-	// Send two requests to test pool reuse
-	for range 2 {
-		req := httptest.NewRequest("GET", "/test", nil)
-		w := httptest.NewRecorder()
-		app.ServeHTTP(w, req)
-		if w.Code != 200 {
-			t.Errorf("status = %d, want 200", w.Code)
-		}
-	}
-}
